Register trip page

+++ db/migrate/20240731153306_create_rounds.rb
@@ -2,7 +2,7 @@ class CreateRounds < ActiveRecord::Migration[5.2]
   def change
     create_table :rounds do |t|
       t.date :date
-      t.float :cost
+      t.integer :cost
       t.references :trip, foreign_key: true
 
       t.timestamps
